Add code for generating rollouts and script that generates them
The rollout code is library code with tests. Ultimately we should
replace this with something from some existing library. I'm taking
the approach in this commit for now in order to decouple progress
on the model from progress on the infrastructure side.

Then there are a couple scripts for generating data and
visualizing it. These don't have tests since they're not part of
the library code. I'll include some visualizations, though.